requirements.txt
albumentations
attrdict
boto3
checksumdir
clip
cython
datasets==2.16.1
decord
deepspeed
decorator
diffusers==0.21.4
easydict
einops
fairscale
ftfy
fvcore
huggingface-hub==0.25.2
jieba
kornia
lpips
matplotlib
more-itertools
ninja
omegaconf
open-clip-torch==2.7.0
openpyxl==3.1.1
opencv-python
pandas
psutil
py-spy
pycocotools
pywavelets
pytorch_lightning
PyYAML
regex
rouge-metric
safetensors
scipy
sentencepiece
streamlit
taming-transformers-rom1504
tensorboard
timm
tqdm
webdataset